add and fix review comments
