+++ Healthy/Classes/Extensions/UILabel+Style/UILabelStyle.swift
@@ -21,3 +21,24 @@ extension UILabel {
         style.applyStyle(for: self)
     }
 }
+
+// MARK:  Title Label Style
+
+
+struct TitleLabelStyle: LabelStyling {
+    func applyStyle(for label: UILabel) {
+        label.textColor = .black20
+        label.font = .poppinsbold(20.0)
+        label.numberOfLines = .zero
+    }
+}
+
+// MARK:  Subtitle Label Style
+
+struct SubtitleLabelStyle: LabelStyling {
+    func applyStyle(for label: UILabel) {
+        label.textColor = .black20
+        label.font = .poppinsbold(18.0)
+        label.numberOfLines = .zero
+    }
+}
